Factor out flowgraph DFS creation
