+++ ILCompiler/Compiler/FlowgraphHelpers/FlowgraphDfsTree.cs
@@ -1,0 +1,55 @@
+﻿namespace ILCompiler.Compiler.FlowgraphHelpers
+{
+    public class FlowgraphDfsTree
+    {
+        public IList<BasicBlock> PostOrder { get; init; }
+
+        public FlowgraphDfsTree(IList<BasicBlock> postOrder)
+        {
+            PostOrder = postOrder;
+        }
+
+        public static FlowgraphDfsTree Build(BasicBlock firstBlock)
+        {
+            var postOrder = new List<BasicBlock>();
+            var visitedBlocks = new HashSet<BasicBlock>();
+
+            var blocksStack = new Stack<BasicBlock>();
+            var successorsStack = new Stack<int>();
+
+            blocksStack.Push(firstBlock);
+            successorsStack.Push(0);
+
+            uint postIndex = 0;
+
+            while (blocksStack.Count > 0)
+            {
+                var topBlock = blocksStack.Peek();
+                var successorIndex = successorsStack.Pop();
+
+                if (successorIndex < topBlock.Successors.Count)
+                {
+                    var successor = topBlock.Successors[successorIndex];
+                    successorsStack.Push(successorIndex + 1);
+
+                    if (!visitedBlocks.Contains(successor))
+                    {
+                        blocksStack.Push(successor);
+                        successorsStack.Push(0);
+
+                        visitedBlocks.Add(successor);
+                    }
+                }
+                else
+                {
+                    var block = blocksStack.Pop();
+                    block.PostOrderNum = postIndex;
+                    postOrder.Add(block);
+                    postIndex++;
+                }
+            }
+
+            return new FlowgraphDfsTree(postOrder);
+        }
+    }
+}
